backend/requirements.txt
playwright
beautifulsoup4
extruct
requests
tldextract
python-Wappalyzer
lxml
fastapi
uvicorn[standard]
matplotlib
jinja2
markdown
